+++ NEsper/NEsper/compat/threading/MonitorSpinLock.cs
@@ -98,7 +98,7 @@ namespace com.espertech.esper.compat.threading
 
         public IDisposable Acquire(long msec)
         {
-            InternalAcquire(msec);
+            InternalAcquire((int) msec);
             return new TrackedDisposable(InternalRelease);
         }
 
@@ -112,7 +112,7 @@ namespace com.espertech.esper.compat.threading
         /// <returns></returns>
         public IDisposable Acquire(bool releaseLock, long? msec = null)
         {
-            InternalAcquire(msec ?? _uLockTimeout);
+            InternalAcquire((int) (msec ?? _uLockTimeout));
             if (releaseLock)
                 return new TrackedDisposable(InternalRelease);
             return new VoidDisposable();
