+++ NEsper/NEsper/compat/threading/StandardReaderWriterLock.cs
@@ -66,11 +66,11 @@ namespace com.espertech.esper.compat.threading
         /// Acquires the reader lock.
         /// </summary>
         /// <param name="timeout">The timeout.</param>
-        public void AcquireReaderLock(int timeout)
+        public void AcquireReaderLock(long timeout)
         {
             try
             {
-                _rwLock.AcquireReaderLock(timeout);
+                _rwLock.AcquireReaderLock((int) timeout);
             }
             catch(ApplicationException)
             {
@@ -82,11 +82,11 @@ namespace com.espertech.esper.compat.threading
         /// Acquires the writer lock.
         /// </summary>
         /// <param name="timeout">The timeout.</param>
-        public void AcquireWriterLock(int timeout)
+        public void AcquireWriterLock(long timeout)
         {
             try
             {
-                _rwLock.AcquireWriterLock(timeout);
+                _rwLock.AcquireWriterLock((int) timeout);
             }
             catch(ApplicationException)
             {
